# backend/public/swagger.json
{
	"openapi": "3.0.0",
	"components": {
...
		"schemas": {
...
			"Record_string.unknown_": {
				"properties": {},
				"additionalProperties": {},
				"type": "object",
				"description": "Construct a type with a set of properties K of type T"
			},
...
			"WorkoutResponse": {
				"description": "Saved workout returned by the API",
				"properties": {
					"id": {
						"type": "string"
					},
					"name": {
						"type": "string"
					},
					"sessionInput": {
						"$ref": "#/components/schemas/Record_string.unknown_"
					},
					"generatedSession": {
						"$ref": "#/components/schemas/Record_string.unknown_"
					},
					"createdAt": {
						"type": "string"
					},
					"updatedAt": {
						"type": "string"
					}
				},
				"required": [
					"id",
					"name",
					"sessionInput",
					"generatedSession",
					"createdAt",
					"updatedAt"
				],
				"type": "object",
				"additionalProperties": false
			},
			"CreateWorkoutBody": {
				"description": "Body for saving a generated workout",
				"properties": {
					"name": {
						"type": "string",
						"description": "Display name for this workout session"
					},
					"sessionInput": {
						"$ref": "#/components/schemas/Record_string.unknown_",
						"description": "The inputs used to generate the session (level, goal, etc.)"
					},
					"generatedSession": {
						"$ref": "#/components/schemas/Record_string.unknown_",
						"description": "The full generated session object"
					}
				},
				"required": [
					"name",
					"sessionInput",
					"generatedSession"
				],
				"type": "object",
				"additionalProperties": false
			},
			"UpdateWorkoutBody": {
				"description": "Body for updating a saved workout",
				"properties": {
					"name": {
						"type": "string",
						"description": "New display name"
					},
					"generatedSession": {
						"$ref": "#/components/schemas/Record_string.unknown_",
						"description": "Updated generated session (when exercises are edited)"
					}
				},
				"type": "object",
				"additionalProperties": false
			},
...
		},
		"securitySchemes": {
			"bearerAuth": {
				"type": "http",
				"scheme": "bearer",
				"bearerFormat": "JWT"
			}
		}
	},
	"info": {
		"title": "GRAVITACIO API",
		"version": "1.0.0",
		"description": "Bouldering workout planner API",
		"contact": {}
	},
	"paths": {
...
		"/api/workouts": {
			"get": {
				"operationId": "ListWorkouts",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"items": {
										"$ref": "#/components/schemas/WorkoutResponse"
									},
									"type": "array"
								}
							}
						}
					}
				},
				"description": "List all saved workouts for the authenticated user, newest first.",
				"tags": [
					"Workouts"
				],
				"security": [
					{
						"bearerAuth": []
					}
				],
				"parameters": []
			},
			"post": {
				"operationId": "CreateWorkout",
				"responses": {
					"201": {
						"description": "Created",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/WorkoutResponse"
								}
							}
						}
					}
				},
				"description": "Save a new generated workout session.",
				"tags": [
					"Workouts"
				],
				"security": [
					{
						"bearerAuth": []
					}
				],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/CreateWorkoutBody"
							}
						}
					}
				}
			}
		},
		"/api/workouts/{id}": {
			"get": {
				"operationId": "GetWorkout",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/WorkoutResponse"
								}
							}
						}
					}
				},
				"description": "Get a single saved workout by ID (must belong to the authenticated user).",
				"tags": [
					"Workouts"
				],
				"security": [
					{
						"bearerAuth": []
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			},
			"put": {
				"operationId": "UpdateWorkout",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/WorkoutResponse"
								}
							}
						}
					}
				},
				"description": "Update a saved workout (name and/or exercises).",
				"tags": [
					"Workouts"
				],
				"security": [
					{
						"bearerAuth": []
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/UpdateWorkoutBody"
							}
						}
					}
				}
			},
			"delete": {
				"operationId": "DeleteWorkout",
				"responses": {
					"204": {
						"description": "No Content"
					}
				},
				"description": "Delete a saved workout.",
				"tags": [
					"Workouts"
				],
				"security": [
					{
						"bearerAuth": []
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
...
	},
	"servers": [
		{
			"url": "/"
		}
	]
}